REconfiguring the db and routes. Hooked up a get all for the static feelings, a get all for the feelings, and a post to feelings.
