+++ controller/static_feelings.js
@@ -1,0 +1,14 @@
+const model = require('../models/static_feelings.js')
+
+
+const getAllFeelings = (req, res, next) => {
+  model.getAllFeelings()
+  .then((result) => {
+    res.send(result)
+  })
+
+}
+
+module.exports = {
+  getAllFeelings
+}
