+++ models/all_feelings.js
@@ -1,0 +1,17 @@
+const knex = require('../knex')
+
+const getFeelings = () => {
+  return knex('feelings')
+}
+
+const createFeeling = (feeling) => {
+  return knex('feelings')
+    .insert(feeling)
+    .then()
+}
+
+
+module.exports = {
+  createFeeling,
+  getFeelings
+}
