+++ models/received_scores.js
@@ -2,7 +2,7 @@ const knex = require('../knex')
 
 
 const getScores = (user_id,is_loved)=> {
-  return knex.select('score', 'feedback', 'description', 'created_at').from('scores')
+  return knex.select('score', 'feedback', 'from_user_id', 'created_at').from('scores')
   .where({to_user_id: user_id})
   .join('users_feelings',`users_feelings.id`, `scores.users_feelings_id`)
   .join('feelings',`users_feelings.feeling_id`, `feelings.id`)
