+++ models/static_feelings.js
@@ -1,0 +1,10 @@
+const knex = require('../knex')
+
+const getAllFeelings = () => {
+  return knex('static_feelings')
+}
+
+
+module.exports = {
+  getAllFeelings
+}
