+++ seeds/006_static_feelings.js
